feat: integrate IconBell component into Navbar and restructure Home page layout
